Graph separation
